test: merge rules/config tests
They are testing the same module.
